extract meshprocessor
